Refactored Data Types added comments. Fixed test cases.
