+++ Library/HL7Enumerator.Core/DataTypes.cs
@@ -2,12 +2,16 @@
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Globalization;
+using System.Linq;
 using System.Text;
 using HL7Enumerator;
 namespace HL7Enumerator.Types
 {
-    public static class DataTypes
+    public static partial class DataTypes
     {
+        private static readonly string[] AllowedComparitors = new string[6] { "<", ">", "=>", "<=", "=", "<>" };
+        private static readonly string[] AllowedSeparators = new string[5] { "-", "+", "/", ".", ":" };
+
         /// <summary>
         /// Converts a DateTime to the LOCAL HL7 DT Format string (YYYYMMDD)  If the Local/Universal Kind 
         /// (DateTime.Kind) property has NOT been set, local Time is assumed.
@@ -140,531 +144,23 @@ namespace HL7Enumerator.Types
                    (idx > 2) ? Constants.MSHSeparators[4] : Constants.MSHSeparators[idx + 1];
         }
 
-        public class EI_EntityIdentifier
-        {
-            public string Identifier { get; set; }
-            public string NamespaceId { get; set; }
-            public string UniversalId { get; set; }
-            public string UniversalIdType { get; set; }
-            public override string ToString()
-            {
-                return ToString('^');
-            }
-            public string ToString(char separator)
-            {
-                return $"{Identifier}{separator}{NamespaceId}{separator}{UniversalId}{separator}{UniversalIdType}";
-            }
-        }
-        public class ED_EncapsulatedData
-        {
-            public HD_HierarchicDesignator SourceApplication;
-            public string TypeOfData;
-            public string DataSybType;
-            public string Encoding;
-            public string Data;
-            public override string ToString()
-            {
-                return ToString('&');
-            }
-            public string ToString(char separator)
-            {
-                return $"{TypeOfData}{separator}{DataSybType}{separator}{Encoding}{separator}{Data}";
-            }
-        }
-
-        public class HD_HierarchicDesignator
-        {
-            public string NamespaceId { get; set; }
-            public string UniversalId { get; set; }
-            public string UniversalIdType { get; set; }
-            public override string ToString()
-            {
-                return ToString('^');
-            }
-            public string ToString(char separator)
-            {
-                return $"{NamespaceId}^{UniversalId}^{UniversalIdType}";
-            }
-        }
-
-        public class CX_CompositeId
-        {
-            public string ID { get; set; }
-            public string CheckDigit { get; set; }
-            public string CheckDigitScheme { get; set; }
-            public HD_HierarchicDesignator AssigningAuthority { get; set; }
-            public string IdentifierTypeCode { get; set; }
-            public HD_HierarchicDesignator AssigningFacility { get; set; }
-            public DateTime? EffectiveDate { get; set; }
-            public DateTime? ExpirationDate { get; set; }
-            public override string ToString()
-            {
-                return ToString('^');
-            }
-            public string ToString(char separator)
-            {
-                var ns = NextSeparator(separator);
-                return
-                    $"{ID}{separator}{CheckDigit}{separator}{CheckDigitScheme}{separator}" +
-                    $"{AssigningAuthority.ToString(ns)}{separator}{AssigningFacility.ToString(ns)}" +
-                    $"{EffectiveDate?.AsDTLocal()}{separator}{ExpirationDate?.AsDTLocal()}";
-            }
-        }
-        public class NM_Number
-        {
-            public NM_Number()
-            {
-            }
-            public NM_Number(HL7Element element)
-            {
-                Value = element;
-            }
-            public NM_Number(long intValue)
-            {
-                Value = intValue.ToString();
-            }
-            public NM_Number(int intValue)
-            {
-                Value = intValue.ToString();
-            }
-            public NM_Number(Double dblValue)
-            {
-                Value = dblValue.ToString();
-            }
-            public NM_Number(float fltValue)
-            {
-                Value = fltValue.ToString();
-            }
-            public NM_Number(decimal decValue)
-            {
-                Value = decValue.ToString();
-            }
-
-            public static implicit operator NM_Number(HL7Element element)
-            {
-                return new NM_Number(element);
-            }
-
-            public string Value { get; set; }
-
-            public static implicit operator double(NM_Number number)
-            {
-                if (number.Value != null && number.Value.Contains("."))
-                {
-                    double newDoubleValue;
-                    return (Double.TryParse(number.Value, out newDoubleValue)) ? newDoubleValue : Double.MinValue;
-                }
-                return Double.MinValue;
-            }
-            public static implicit operator long(NM_Number number)
-            {
-                if (number.Value == null) return long.MinValue;
-                long newInt64Value;
-                return (long.TryParse(number.Value, out newInt64Value)) ? newInt64Value : long.MinValue;
-
-            }
-            public static implicit operator int(NM_Number number)
-            {
-                if (number.Value == null) return int.MinValue;
-                int newIntValue;
-                return (int.TryParse(number.Value, out newIntValue)) ? newIntValue : int.MinValue;
-            }
-            public static implicit operator decimal(NM_Number number)
-            {
-                if (number.Value == null) return decimal.MinValue;
-                decimal newDecValue;
-                return (Decimal.TryParse(number.Value, out newDecValue)) ? newDecValue : Decimal.MinValue;
-            }
-            public static implicit operator string(NM_Number number)
-            {
-                if (number.Value == null) return String.Empty;
-
-                string value = number.Value;
-                if (value.Contains("."))
-                {
-                    double newDoubleValue;
-                    return (Double.TryParse(value, out newDoubleValue)) ? newDoubleValue.ToString() : String.Empty;
-                }
-                else
-                {
-                    long newInt64Value;
-                    return (Int64.TryParse(value, out newInt64Value)) ? newInt64Value.ToString() : String.Empty;
-                }
-            }
-        }
-        public class SN_Number : NM_Number
-        {
-            public string Comparitor { get; set; }
-            public NM_Number Num1 { get; set; }
-            public string SeparatorOrSuffix { get; set; }
-            public NM_Number Num2 { get; set; }
-        }
-
-        public class XON_ExtendedCompositeNameForOrganizations 
-        {   
-            public string OrganizationName { get; set; }
-            public string OrganizationNameTypeCode { get; set; }
-            public NM_Number ID { get; set; }
-            public string CheckDigit { get; set; }
-            public string CheckDigitScheme { get; set; }
-            public HD_HierarchicDesignator AssigningAuthority { get; set; }
-            public string IdentifierTypeCode { get; set; }
-            public HD_HierarchicDesignator AssigningFacility { get; set; }
-        }
-        public class XPN_ExtendedPersonName
-        {
-            public string FamilyName { get; set; }
-            public string GivenName { get; set; }
-            public string SecondGivenNamesOrInitials { get; set; }
-            public string Suffix { get; set; }
-            public string Prefix { get; set; }
-            public string Degree { get; set; }
-            public string NameTypeCode { get; set; }
-            public string NameRepresentationCode { get; set; }
-            public CE_CodedElement NameContext { get; set; }
-            public string NameAssemblyOrder { get; set; }
-            public override string ToString()
-            {
-                return ToString('^');
-            }
-            public virtual string ToString(char separator)
-            {
-                var ns = NextSeparator(separator);
-                return
-                    $"{FamilyName}{separator}{GivenName}{separator}{SecondGivenNamesOrInitials}" +
-                    $"{separator}{Suffix}{separator}{Prefix}{separator}{Degree}{separator}" +
-                    $"{NameTypeCode}{separator}{NameRepresentationCode}{separator}"+
-                    $"{NameContext.ToString(ns)}{separator}{NameAssemblyOrder}";
-            }
-        }
-
-        public class TN_TelephoneNumber
-        {
-            public string CountryCode { get; set; }
-            public string AreaCode { get; set; }
-            public string Number { get; set; }
-            public string Extension { get; set; }
-            public string Beeper { get; set; }
-            public string Comment { get; set; }
-            public TN_TelephoneNumber()
-            { 
-
-            }
-
-            public TN_TelephoneNumber(HL7Element element)
-            {
-                this.Parse((string)element);
-            }
-            private enum TNState {None, CountryCode, AreaCode, Number, Extension, Beeper, Comment};
-            public static bool TryParse(string tn, out TN_TelephoneNumber tnNumber)
-            {
-                tnNumber = null;
-                var number = new TN_TelephoneNumber();
-                if (number.Parse(tn, false))
-                {
-                    tnNumber = number;
-                    return true;
-                }
-                else return false;
-            }
-            public bool Parse(string tn, bool throwException=true)
-            {
-                if (string.IsNullOrWhiteSpace(tn))
-                {
-                    if (throwException) throw new FormatException("Invalid TN Format");
-                    else return false;
-                }
-                var done = false;
-                var state = TNState.None;
-                var index = 0;
-                while (index < tn.Length && !done)
-                {
-                    char c = tn[index++];
-                    switch (state)
-                    {
-                        case TNState.None:
-                            if (Char.IsDigit(c)) state = TNState.CountryCode;
-                            else if (c.Equals('(')) state = TNState.AreaCode;
-                            break;
-                        case TNState.CountryCode:
-                            if (!char.IsDigit(c)) state = TNState.AreaCode;
-                            else CountryCode = $"{CountryCode}{c}";
-                            if (CountryCode.Length > 2)
-                            {
-                                Number = CountryCode;
-                                CountryCode = "";
-                                AreaCode = "";
-                                state = TNState.Number;
-                            }
-                            break;
-                        case TNState.AreaCode:
-                            if (!c.Equals('('))
-                            {
-                                if (!char.IsDigit(c)) state = TNState.Number;
-                                AreaCode = $"{AreaCode}{c}";
-                                if (AreaCode.Length > 3)
-                                {
-                                    Number = AreaCode;
-                                    AreaCode = "";
-                                }
-                            }
-                            break;
-                        case TNState.Number:
-                            if (char.IsDigit(c) || c.Equals('-'))
-                            {
-                                Number = $"{Number}{c}";
-                            }
-                            else
-                            {
-                                switch (c)
-                                {
-                                    case ' ':
-                                        break; // ignore these
-                                    case 'X':
-                                        state = TNState.Extension;
-                                        break;
-                                    case 'B':
-                                        state = TNState.Beeper;
-                                        break;
-                                    case 'C':
-                                        state = TNState.Comment;
-                                        break;
-                                    default:
-                                        if (throwException) throw new FormatException($"Invalid TN Format Number '{c}' not expected at {1+index}");
-                                        else return false;
-                                }
-                            }
-                            break;
-                        case TNState.Extension:
-                            if (char.IsDigit(c) || c.Equals('-'))
-                            {
-                                Extension = $"{Extension}{c}";
-                            }
-                            else
-                            {
-                                switch (c)
-                                {
-                                    case ' ':
-                                        break; // ignore these
-                                    case 'B':
-                                        state = TNState.Beeper;
-                                        break;
-                                    case 'C':
-                                        state = TNState.Comment;
-                                        break;
-                                    default:
-                                        if (throwException) throw new FormatException($"Invalid TN Format Extension '{c}' not expected at {1 + index}");
-                                        else return false;
-                                }
-                            }
-                            break;
-                        case TNState.Beeper:
-                            if (char.IsDigit(c) || c.Equals('-'))
-                            {
-                                Beeper = $"{Extension}{c}";
-                            }
-                            else
-                            {
-                                switch (c)
-                                {
-                                    case ' ':
-                                        break; // ignore these
-                                    case 'C':
-                                        state = TNState.Comment;
-                                        break;
-                                    default:
-                                        if (throwException) throw new FormatException($"Invalid TN Format Beeper '{c}' not expected at {1 + index}");
-                                        else return false;
-                                }
-                            }
-                            break;
-                        case TNState.Comment:
-                            Comment = tn.Substring(index);
-                            done = true;
-                            break;
-                    }
-
-                }
-                return true;
-            }
-            public override string ToString()
-            {
-                var tn = (CountryCode==null) ? $"{CountryCode} ": "";
-                if (AreaCode != null) tn = $"{tn}({AreaCode})";
-                tn = $"{tn}{Number}";
-                if (Extension != null) tn = $"{tn}X{Extension}";
-                if (Beeper != null) tn = $"{tn}B{Beeper}";
-                if (Comment != null) tn = $"{tn}C{Comment}";
-                return tn;
-            }
-
-            public static implicit operator string(TN_TelephoneNumber number)
-            {
-                return (number==null) ? String.Empty : number.ToString();
-            }
-            public static implicit operator TN_TelephoneNumber(string tn)
-            {
-                var tnNumber = new TN_TelephoneNumber();
-                tnNumber.Parse(tn);
-                return tnNumber;
-            }
-        }
-        public class XTN_ExtendedTelecommunicationNumber
-        {
-            public TN_TelephoneNumber TelephoneNumber { get; set; }
-            public string TelecommunicationUseCode { get; set; }
-            public string TelecommunicationEquipment { get; set; }
-            public string EmailAddress { get; set; }
-            public override string ToString()
-            {
-                return ToString('^');
-            }
-            public virtual string ToString(char separator)
-            {
-                return
-                    $"{(string)TelephoneNumber}{separator}{TelecommunicationUseCode}{separator}{TelecommunicationEquipment}" +
-                    $"{separator}{EmailAddress}";
-            }
-        }
-
-        public class XCN_ExtendedCompositeIDAndName:XPN_ExtendedPersonName
-        {
-            public string ID { get; set; }
-            public string SourceTable { get; set; }
-            public HD_HierarchicDesignator AssigningAuthority { get; set; }
-            public string IdentifierCheckDigit { get; set; }
-            public string CheckDigitScheme { get; set; }
-            public string IdentifierTypeCode { get; set; }
-            public HD_HierarchicDesignator AssigningFacility { get; set; }
-            public DR_DateRange NameValidityRange { get; set; }
-            public override string ToString(char separator)
-            {
-                var ns = NextSeparator(separator);
-                return
-                    $"{ID}{separator}{FamilyName}{separator}{GivenName}{separator}{SecondGivenNamesOrInitials}" +
-                    $"{separator}{Suffix}{separator}{Prefix}{separator}{Degree}{separator}" +
-                    $"{SourceTable}{separator}{AssigningAuthority.ToString(ns)}{separator}" +
-                    $"{NameTypeCode}{separator}{IdentifierCheckDigit}{separator}{CheckDigitScheme}{separator}" +
-                    $"{IdentifierTypeCode}{separator}{AssigningFacility.ToString(ns)}{separator}" +
-                    $"{NameRepresentationCode}{separator}{NameContext.ToString(ns)}{separator}" +
-                    $"{NameValidityRange.ToString(ns)}{separator}{NameAssemblyOrder}";
-            }
-        }
-        public class VID_VersionIdentifier
-        { 
-           public string VersionID { get; set; }
-           public CE_CodedElement Internationalization { get; set; }
-           public CE_CodedElement InternationalVersionID { get; set; }
-            public override string ToString()
-            {
-                return ToString('^');
-            }
-            public virtual string ToString(char separator)
-            {
-                return
-                    $"{VersionID}{separator}{Internationalization.ToString()}{separator}{InternationalVersionID.ToString()}";
-            }
-
-        }
-        public class DR_DateRange
-        {
-            public DateTime? DateFrom { get; set; }
-            public DateTime? DateTo { get; set; }
-            public override string ToString()
-            {
-                return ToString('^');
-            }
-            public string ToString(char separator)
-            {
-                return $"{DateFrom?.AsTSLocal()}{separator}{DateTo?.AsTSLocal()}";
-            }
-        }
-
-        public class CE_CodedElement
-        {
-            public string Identifier { get; set; }
-            public string Text { get; set; }
-            public string NameOfCodingSystem { get; set; }
-            public string AlternateIdentifier { get; set; }
-            public string AlternateText { get; set; }
-            public string NameOfAlternateCodingSystem { get; set; }
-            public override string ToString()
-            {
-                return ToString('^');
-            }
-            public virtual string ToString(char sepatator)
-            {
-                return $"{Identifier}{sepatator}{Text}{sepatator}{NameOfCodingSystem}" +
-                    $"{sepatator}{AlternateIdentifier}{sepatator}{AlternateText}" +
-                    $"{sepatator}{NameOfAlternateCodingSystem}";
-            }
-        }
-
-
-        public class CNE_CodedWithNoExceptions : CE_CodedElement
-        {
-            public string CodingSystemVersionID { get; set; }
-            public string AlternateCodingSystemVersionID { get; set; }
-            public string OriginalText { get; set; }
-            public override string ToString(char separator)
-            {
-               return $"{base.ToString(separator)}{separator}{CodingSystemVersionID}{separator}" +
-                    $"{AlternateCodingSystemVersionID}{separator}{OriginalText}";
-            }
-        }
-        public class CWE_CodedWithExceptions : CNE_CodedWithNoExceptions
-        { 
-        }
-
-        public class SAD_StreetAddress
-        {
-            public string StreetOrMailingAddress { get; set; }
-            public string StreetName { get; set; }
-            public string DwellingNumber { get; set; }
-            public override string ToString()
-            {
-                return ToString('^');
-            }
-            public string ToString(char sepatator)
-            {
-                return $"{StreetOrMailingAddress}{sepatator}{StreetName}{sepatator}{DwellingNumber}";
-            }
-        }
-
-        public class XAD_ExtendedAddress
-        {
-            public SAD_StreetAddress StreetAddress { get; set; }
-            public string OtherDesignation { get; set; }
-            public string City { get; set; }
-            public string StateOrProvince { get; set; }
-            public string ZipOrPostalCode { get; set; }
-            public string Country { get; set; }
-            public string AddressType { get; set; }
-            public string OtherGeographicDesignation { get; set; }
-            public string CountyOrParishCode { get; set; }
-            public string CensusTract { get; set; }
-            public string AddressRepresentationCode { get; set; }
-            public DR_DateRange AddressValidityRange { get; set; }
-            public override string ToString()
-            {
-                return ToString('^');
-            }
-            public string ToString(char sepatator)
-            {
-                return $"{StreetAddress.ToString()}{sepatator}{OtherDesignation}{sepatator}{City}" +
-                    $"{sepatator}{StateOrProvince}{sepatator}{ ZipOrPostalCode}" +
-                    $"{sepatator}{Country}{sepatator}{AddressType}{sepatator}{OtherGeographicDesignation}" +
-                    $"{sepatator}{CountyOrParishCode}{sepatator}{CensusTract}{sepatator}{AddressValidityRange.ToString()}";
-            }
-        }
-
-
-
+        /// <summary>
+        /// Safely locate the desired element If it exists (returning defaults)
+        /// </summary>
+        /// <param name="element">HL7 Element</param>
+        /// <param name="index">the index in the element required</param>
+        /// <returns>A HL7 element if it exists</returns>
         public static HL7Element IndexedElement(this HL7Element element, int index = -1)
         {
             return (index == -1) ? element : (index < element.Count) ? element[index] : null;
         }
-
+        /// <summary>
+        /// Safely return Value of the desired Element.  This relieves the need to check 
+        /// for ethe existence or null value of the required element.
+        /// </summary>
+        /// <param name="element">The Containing element element</param>
+        /// <param name="index">Zero based index of the element required</param>
+        /// <returns>A string represenation of the hl7 field.</returns>
         public static string ElementValue(this HL7Element element, int index)
         {
             return (element == null) ? null :
@@ -672,19 +168,24 @@ namespace HL7Enumerator.Types
                    (index == 0 && element.Count == 0) ? element.Value :
                    (element.Count > index) ? element[index].ToString() : "";
         }
-
+        /// <summary>
+        /// Safely Extract the Supplied element assuming the element contains suitable contents.
+        /// </summary>
+        /// <param name="element">Element representing or containing the HD Data</param>
+        /// <param name="index">Optionally the sub element </param>
+        /// <returns>A fully populated HL7 HD object</returns>
         public static HD_HierarchicDesignator AsHD(this HL7Element element, int index = -1)
         {
             var el = element.IndexedElement(index);
             if (el == null) return null;
-            return new HD_HierarchicDesignator()
-            {
-                NamespaceId = (el.Count == 0) ? el.Value : el[0].Value,
-                UniversalId = (el.Count < 2) ? "" : el[1].Value,
-                UniversalIdType = (el.Count < 3) ? "" : el[2].Value
-            };
+            return new HD_HierarchicDesignator(el);
         }
-
+        /// <summary>
+        /// Safely Extract a Date Range from the Supplied element assuming suitable contents;
+        /// </summary>
+        /// <param name="element"></param>
+        /// <param name="index"></param>
+        /// <returns></returns>
         public static DR_DateRange AsDateRange(this HL7Element element, int index = -1)
         {
             var el = element.IndexedElement(index);
@@ -695,17 +196,43 @@ namespace HL7Enumerator.Types
                 DateTo = (el.Count < 2) ? ToLocalDateTime("") : ToLocalDateTime(el[1].Value)
             };
         }
+        /// <summary>
+        /// Extract a Date Time field from a supplied element assuming suitable content
+        /// NOTE: Dates use the DateTime.Kind property to correctly convert the date
+        /// from UTC and Local Times.  Use the LocaLtime and UTCTime extensions to safely
+        /// handle the returned date.
+        /// </summary>
+        /// <param name="element"></param>
+        /// <param name="index"></param>
+        /// <returns>A Date time as UTC or Local Time depending on specificity of field data</returns>
         public static DateTime? FromDT(this HL7Element element, int index = -1)
         {
             var dtTm = element.FromTS(index);
             return dtTm?.Date;
         }
+        /// <summary>
+        /// Extract a Time field from a supplied element assuming suitable content
+        /// NOTE: Dates and Tims use the DateTime.Kind property to correctly convert the date
+        /// from UTC and Local Times.  Use the LocaLtime and UTCTime extensions to safely
+        /// handle the returned Time.
+        /// </summary>
+        /// <param name="element"></param>
+        /// <param name="index"></param>
+        /// <returns></returns>
         public static TimeSpan? FromTM(this HL7Element element, int index = -1)
         {
             var dtTm = AsDateTime($"20000101{element.ElementValue(index)}");
             return dtTm?.TimeOfDay;
         }
-
+        /// <summary>
+        /// Extract a DateTime field from a supplied element assuming suitable content
+        /// NOTE: Dates and Tims use the DateTime.Kind property to correctly convert the date
+        /// from UTC and Local Times.  Use the LocaLtime and UTCTime extensions to safely
+        /// handle the returned DateTime.
+        /// </summary>
+        /// <param name="element"></param>
+        /// <param name="index"></param>
+        /// <returns></returns>
         public static DateTime? FromTS(this HL7Element element, int index = -1)
         {
             var el = element.ElementValue(index);
@@ -719,7 +246,16 @@ namespace HL7Enumerator.Types
                 return null;
             }
         }
-
+        /// <summary>
+        /// Safely return a DataTime from a HL7 Timestamp. NOTE: It is not usually necessary
+        /// to call this method directly. Use the ExtractTimeZone method 
+        /// 
+        /// Dates and Tims use the DateTime.Kind property to correctly convert the date
+        /// from UTC and Local Times.  Use the LocaLtime and UTCTime extensions to safely
+        /// handle the returned DateTime.
+        /// </summary>
+        /// <param name="hl7TS"></param>
+        /// <returns></returns>
         public static DateTime? AsDateTime(string hl7TS)
         {
             var tsDt = ExtractTimeZone(hl7TS, out string zone);
@@ -729,7 +265,12 @@ namespace HL7Enumerator.Types
             if (zone?.Length > 0) return DateTime.Parse(dtText, CultureInfo.InvariantCulture, DateTimeStyles.AdjustToUniversal);
             return DateTime.Parse(dtText, CultureInfo.InvariantCulture, DateTimeStyles.AssumeLocal);
         }
-
+        /// <summary>
+        /// Converts a HL7 Text string to ISO 8601 format
+        /// </summary>
+        /// <param name="dateText"></param>
+        /// <param name="zone"></param>
+        /// <returns></returns>
         public static string HL7DateTextAsISODateText(string dateText, string zone = null)
         {
             // Date And Time (DT/TM) are a subset of HL7 Timestamps(TS) 
@@ -744,7 +285,6 @@ namespace HL7Enumerator.Types
                 dateText = ExtractTimeZone(dateText, out zone);
                 if (!Double.TryParse(dateText, out _)) return null;
             };
-
 
             var iso = "";
             if (dateText.Length < 4) return iso;
@@ -809,36 +349,13 @@ namespace HL7Enumerator.Types
             tzText = tz;
             return dateText;
         }
-
-        public static CX_CompositeId AsCX(this HL7Element element)
-        {
-            return new CX_CompositeId()
-            {
-                ID = element.ElementValue(0),
-                CheckDigit = element.ElementValue(1),
-                CheckDigitScheme = element.ElementValue(2),
-                AssigningAuthority = element.AsHD(3),
-                IdentifierTypeCode = element.ElementValue(4),
-                AssigningFacility = element.AsHD(5),
-                EffectiveDate = element.FromTS(6),
-                ExpirationDate = element.FromTS(7),
-            };
-        }
-        public static ED_EncapsulatedData AsED(this HL7Element element)
-        {
-            if (element.Count == 0)
-                return new ED_EncapsulatedData() { Data = element };
-
-            return new ED_EncapsulatedData()
-            {
-                SourceApplication = element.AsHD(0),
-                TypeOfData = element.ElementValue(1),
-                DataSybType = element.ElementValue(2),
-                Encoding = element.ElementValue(3),
-                Data = element.ElementValue(4)
-            };
-        }
-
+        /// <summary>
+        /// Return the LOCAL time from a HL7 Encoded string;
+        /// NOTE: the DateTime.Kind property will be set appropriately depending on the 
+        /// specificity of the TimeStamp String.
+        /// </summary>
+        /// <param name="value"></param>
+        /// <returns></returns>
         public static DateTime? ToLocalDateTime(string value)
         {
             var dt = AsDateTime(value);
@@ -846,6 +363,13 @@ namespace HL7Enumerator.Types
             if (dt.Value.Kind == DateTimeKind.Unspecified) return DateTime.SpecifyKind(dt.Value, DateTimeKind.Local);
             return dt.Value.ToLocalTime();
         }
+        /// <summary>
+        /// Returns the UTC time from a HL7 Encoded string 
+        /// NOTE: the DateTime.Kind property will be set appropriately depending on the 
+        /// specificity of the TimeStamp String.
+        /// </summary>
+        /// <param name="value"></param>
+        /// <returns></returns>
         public static DateTime? ToUTCDateTime(string value)
         {
             var dt = AsDateTime(value);
@@ -857,11 +381,21 @@ namespace HL7Enumerator.Types
         {
             return value.Kind != DateTimeKind.Utc;
         }
+        /// <summary>
+        /// Determine if a HL7 Encoded string represent Local Time.
+        /// </summary>
+        /// <param name="value"></param>
+        /// <returns></returns>
         public static bool IsLocalTime(string value)
         {
             var tm = AsDateTime(value);
             return (tm == null) ? true : IsLocalTime(tm.Value);
         }
+        /// <summary>
+        /// Encapsulates the Int Try/Parse for a string value
+        /// </summary>
+        /// <param name="value"></param>
+        /// <returns></returns>
         private static int AsInteger(string value)
         {
             if (Int32.TryParse(value, out int intValue))
@@ -870,51 +404,97 @@ namespace HL7Enumerator.Types
             }
             return 0;
         }
-
+        /// <summary>
+        /// Automatically convert a HL7 NM to an integer (If possible)
+        /// </summary>
+        /// <param name="number"></param>
+        /// <returns>Integer value or Int.Minimum if not convertable</returns>
+        public static int AsInteger(this NM_Number number)
+        {
+            return (int)number;
+        }
+        /// <summary>
+        /// Automatically convert a HL7 NM to Double value (if possible)
+        /// </summary>
+        /// <param name="number"></param>
+        /// <returns>double value or Double.Minimum if not convertable</returns>
+        public static double AsDouble(this NM_Number number)
+        {
+            return (double)number;
+        }
+        /// <summary>
+        /// Automatically convert a HL7 NM to a Single (Float) value (if possible)
+        /// </summary>
+        /// <param name="number"></param>
+        /// <returns>A Float Value or Float.Minimum if not convertable</returns>
+        public static float AsFloat(this NM_Number number)
+        {
+            return (float)number;
+        }
+        /// <summary>
+        /// Automatically convert a HL7 NM to a Decimal value (if possible)
+        /// </summary>
+        /// <param name="number"></param>
+        /// <returns>a Decimal value or Decimal.Minimum if not convertable</returns>
+        public static decimal AsDecimal(this NM_Number number)
+        {
+            return (decimal)number;
+        }
+        /// <summary>
+        /// Automatically convert a HL7 NM to a 64 bit Integer value (if possible)
+        /// </summary>
+        /// <param name="number"></param>
+        /// <returns>An Int64 value or Int64.Minimum if not convertable</returns>
+        public static Int64 AsInt64(this NM_Number number)
+        {
+            return (Int64)number;
+        }
+        /// <summary>
+        /// Safely Extract a HL7 SN (structured Numeric) from a HL7 Element assuming suitable content
+        /// </summary>
+        /// <param name="element"></param>
+        /// <returns>a SN_Structured Numeric</returns>
+        public static SN_StructuredNumeric AsSN(this HL7Element element)
+        {
+            return new DataTypes.SN_StructuredNumeric(element);
+        }
+        /// <summary>
+        /// Safely Extract a HL7 CX (Composite ID) from a HL7 Element assuming suitable content
+        /// </summary>
+        /// <param name="element"></param>
+        /// <returns>A HL7 CX object</returns>
+        public static CX_CompositeId AsCX(this HL7Element element)
+        {
+            return new CX_CompositeId(element);
+        }
+        /// <summary>
+        /// Safely Extract a ED (Encapsulated Data) from a HL7 Element Assuming suitable content (eg HL7 (ED) OBX[5])
+        /// </summary>
+        /// <param name="element"></param>
+        /// <returns>a HL7 ED object</returns>
+        public static ED_EncapsulatedData AsED(this HL7Element element)
+        {
+            return new ED_EncapsulatedData(element);
+        }
+        /// <summary>
+        /// Safely Extract a HL7 XCN (Extended Composite ID and Name) from a HL7 Element Assuming suitable conent
+        /// </summary>
+        /// <param name="element"></param>
+        /// <returns>A HL7 XCN Object</returns>
         public static XCN_ExtendedCompositeIDAndName AsXCN(this HL7Element element)
         {
-            if (element.Separator == '~')
-            {
-                throw new InvalidOperationException("AsXCN() called on repeating field.\r\n" +
-                    " The Field in question should be treated as an Enumerable type");
-            }
-            return new XCN_ExtendedCompositeIDAndName()
-            {
-                ID = element.ElementValue(0),
-                FamilyName = element.ElementValue(1),
-                GivenName = element.ElementValue(2),
-                SecondGivenNamesOrInitials = element.ElementValue(3),
-                Suffix = element.ElementValue(4),
-                Prefix = element.ElementValue(5),
-                Degree = element.ElementValue(6),
-                SourceTable = element.ElementValue(7),
-                AssigningAuthority = element.IndexedElement(8).AsHD(),
-                NameTypeCode = element.ElementValue(9),
-                IdentifierCheckDigit = element.ElementValue(10),
-                CheckDigitScheme = element.ElementValue(11),
-                IdentifierTypeCode = element.ElementValue(12),
-                AssigningFacility = element.IndexedElement(13).AsHD(),
-                NameRepresentationCode = element.ElementValue(14),
-                NameContext = element.IndexedElement(15).AsCE(),
-                NameValidityRange = element.IndexedElement(16).AsDateRange(),
-                NameAssemblyOrder = element.ElementValue(0)
-            };
+            return new XCN_ExtendedCompositeIDAndName(element);
         }
-
+        /// <summary>
+        /// Safely Extract a HL7 CE Coded Element type from a HL7 Element assuming sutiable content
+        /// </summary>
+        /// <param name="element"></param>
+        /// <returns>a HL7 CE object</returns>
         public static CE_CodedElement AsCE(this HL7Element element)
         {
-            return new CE_CodedElement()
-            {
-                Identifier = element.ElementValue(0),
-                Text = element.ElementValue(1),
-                NameOfCodingSystem = element.ElementValue(2),
-                AlternateIdentifier = element.ElementValue(3),
-                AlternateText = element.ElementValue(4),
-                NameOfAlternateCodingSystem = element.ElementValue(5)
-            };
+            return new CE_CodedElement(element);
 
         }
-
     }
 }
