+++ Library/HL7Enumerator.Core/HL7Types/DataTypes.CWE_CodedWithExceptions.cs
@@ -1,0 +1,11 @@
+﻿namespace HL7Enumerator.Types
+{
+    public static partial class DataTypes
+    {
+        public class CWE_CodedWithExceptions : CNE_CodedWithNoExceptions
+        { 
+        }
+
+    }
+}
+
